Improve DWIM buffer killing
`kill-this-buffer' is meant for use from the menu bar;
`kill-current-buffer' is the generic alternative. Replace usage of
`kill-this-buffer' and update the DWIM wrapper name accordingly.

The prefix argument to `server-edit' just starts the server if it hasn’t
been started yet, which isn’t very useful from a `kill-current-buffer'
wrapper, so don’t pass through the wrapper prefix argument. Instead, the
prefix argument now calls `kill-buffer' interactively to prompt for a
buffer, even in server buffers. Note that server.el adds
`server-kill-buffer' to `kill-buffer-hook', so killing server buffers
with `kill-buffer' still tells the client the buffer is done, although
the next server editing buffer won’t be made active.

Check if the server is running with `server-process' instead of
`daemonp` since daemonp is set only when the server is started via the
command line (as opposed to via `server-mode').