Add poi, remove my/next-error-delegate
poi is a package for moving to the next point of interest.

You can add commands that moves the cursor to a point of
interest (like `jinx-next' and `flymake-goto-next-error') to
`poi-functions'.

Then `poi-next' and `poi-previous' will let you move to the next
point of interest (poi), and `poi-action' will perform a logical
action associated with that point (like `jinx-correct').

This makes it easy to step through points of interest accumulated
from different sources with a single command.

You can go further and integrate this with `next-error' using

  (advice-add 'next-error :around #'poi-delegate)

Now `next-error' will run if its source buffer is visible on
screen, otherwise it will delegate to `poi-next'/`poi-previous'.

If you want to add to `poi-functions' only when turning on
relevant modes, you can use `poi-register':

  (add-hook 'jinx-mode-hook
            (poi-register 'jinx-next 'jinx-correct))

-----

Remove my/next-error-delegate and friends, longstanding hacks in
the config, as this functionality was moved into poi and cleaned
up.